add[BE]: Add ChannelClass.ts
