+++ backend/src/api/channels/channels.service.ts
@@ -4,50 +4,44 @@ import { WebSocketServer } from '@nestjs/websockets';
 import { CreateChannelDto } from './dtos/channel.dto';
 import { ONESECOND } from '../../const';
 
+import { ChannelClass, Channel } from './ChannelClass';
+
 type userId = number;
 type channelId = number;
 type messageId = number;
 
-export interface Channel {
-  id: number;
-  title?: string;
-  isDm: boolean;
-  isPrivate: boolean;
-  createdAt: Date;
-  password?: string;
-
-  owner?: userId;
-  users: Set<userId>;
-  admin: Set<userId>;
-  banned: Set<userId>;
-  muted: Map<userId, Date>;
-}
-
 export interface ChannelUser {
   id: number;
-  name: string;
-  socket;
+  name: string;               // nickname: string
 
-  joined: Set<channelId>;
-  invited: Set<channelId>;
-  blockUser: Set<channelId>;
+  joined: Set<channelId>;     // channels: channel
+  blockUser: Set<channelId>;  // blockeds: user
+
+  socket;
 }
 
 export interface Message {
   id: number;
   senderId: userId;
   channelId: userId;
-  isLog: boolean;
   text: string;
   createdAt: Date;
+
+  isLog: boolean;             // none
+
+  // sender: user;
+  // channel: channel;
 }
 
 @Injectable()
 export class ChannelsService {
   @WebSocketServer() server: Server;
-  private channelMap = new Map<channelId, Channel>();
   private channelUserMap = new Map<userId, ChannelUser>();
   private messageMap = new Map<messageId, Message>();
+
+  public channelClass: ChannelClass = new ChannelClass();
+
+  // User getter
 
   hasUser(userId: number): boolean {
     return this.channelUserMap.has(userId);
@@ -64,10 +58,6 @@ export class ChannelsService {
     return user;
   }
 
-  setUser(userId: number, user: ChannelUser): void {
-    this.channelUserMap.set(userId, user);
-  }
-
   getUserFromSocket(socket: Socket) {
     if (!socket || !socket.data || !socket.data.user) {
       return null;
@@ -80,6 +70,14 @@ export class ChannelsService {
     return [...this.channelUserMap.keys()];
   }
 
+
+
+  // User setter
+
+  setUser(userId: number, user: ChannelUser): void {
+    this.channelUserMap.set(userId, user);
+  }
+
   setUserLeave(user: ChannelUser, channel: Channel) {
     user.socket.leave(String(channel.id));
     user.joined.delete(channel.id);
@@ -87,23 +85,14 @@ export class ChannelsService {
     channel.admin.delete(user.id);
   }
 
-  getChannel(channelId: number) {
-    const channel = this.channelMap.get(channelId);
 
-    if (!channel) {
-      const code = HttpStatus.BAD_REQUEST;
-      const message = 'This channel does not exist.';
-      throw { code, message };
-    } 
-    return channel;
-  }
 
-  getChannelValues() {
-    return [...this.channelMap.values()];
-  }
 
+  // Channel getter
+
+  // 채널의 상세 정보를 찾는다.
   getUserJoinedChannel(userId: number, channelId: number): ChannelUser {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const user: ChannelUser = this.getUser(userId);
 
     if (!channel.users.has(user.id)) {
@@ -114,8 +103,9 @@ export class ChannelsService {
     return user;
   }
 
+  // 채널 목록을 찾는다.
   getUserArrayJoinedChannel(channelId: number) {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const data = [];
 
     channel.users.forEach((userId) => {
@@ -131,8 +121,9 @@ export class ChannelsService {
     return data;
   }
 
+  // 채널에서 유저의 역할을 가져온다.
   getChannelUserRole(channelId: number, userId: number): string {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const user = this.getUserJoinedChannel(userId, channelId);
 
     if (channel.owner == user.id) {
@@ -143,8 +134,51 @@ export class ChannelsService {
     return 'normal';
   }
 
+
+  // Channel setter
+
+  
+
+  
+
+  
+
+  kick(user: ChannelUser, channel: Channel, kickedUser: ChannelUser) {
+    this.setUserLeave(kickedUser, channel);
+    this.noticeToChannel(
+      channel.id,
+      `${user.name}님이 ${kickedUser.name} 님을 채널에서 강퇴하였습니다.`
+    );
+  }
+
+  ban(user: ChannelUser, channel: Channel, bannedUser: ChannelUser) {
+    this.setUserLeave(bannedUser, channel);
+    channel.banned.add(bannedUser.id);
+    this.noticeToChannel(
+      channel.id,
+      `${user.name}님이 ${bannedUser.name} 님을 채널에서 차단하였습니다.`
+    );
+  }
+
+  mute(user: ChannelUser, channel: Channel, mutedUser: ChannelUser, seconds: number) {
+    const now = new Date();
+    const expiresAt = new Date(now.getTime() + (ONESECOND * seconds));
+
+    channel.muted.set(mutedUser.id, expiresAt);
+    this.noticeToChannel(
+      channel.id,
+      `${user.name}님이 ${mutedUser.name} 님을 채널에서 조용히 시켰습니다.`
+    );
+  }
+
+
+
+
+  // Message
+
+  // 채널에 공지를 보낸다.
   noticeToChannel(channelId: number, message: string): Message {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
 
     const newMessage: Message = {
       id: this.messageMap.size + 1,
@@ -166,23 +200,19 @@ export class ChannelsService {
     return newMessage;
   }
 
+  // 채널에 메세지를 보낸다.
   messageToChannel(
     userId: number,
     channelId: number,
     message: string
   ): Message {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const user: ChannelUser = this.getUserJoinedChannel(userId, channelId);
 
     if (channel.muted.has(user.id)) {
       const expiresAt = new Date(channel.muted.get(user.id));
 
-      console.log('muted', expiresAt <= new Date());
       if (expiresAt.getTime() <= new Date().getTime()) {
-        
-        console.log('expired', expiresAt.getTime());
-        console.log('now', new Date().getTime());
-
         channel.muted.delete(user.id);
       } else {
         const code = HttpStatus.FORBIDDEN;
@@ -215,12 +245,13 @@ export class ChannelsService {
     return newMessage;
   }
 
+  // Message getter
+
   getChannelMessages(userId: number, channelId: number) {
-    this.channelMap.get(channelId);
     this.getUserJoinedChannel(userId, channelId);
 
     const data = [];
-    [...this.messageMap.values()].forEach((message) => {
+    [ ...this.messageMap.values() ].forEach((message) => {
       if (message.channelId == channelId) {
         const tarMessage = {
           id: message.id,
@@ -242,39 +273,22 @@ export class ChannelsService {
     return data;
   }
 
-  createChannel(data) {
-    const newChannelId: number = this.channelMap.size + 1;
-    const newChannel: Channel = {
-      id: newChannelId,
-      title: data.title ? data.title : null,
-      isDm: data.isDm ? true : false,
-      isPrivate: data.isPrivate ? true : false,
-      createdAt: new Date(),
-      password: data.password ? data.password : null,
 
-      owner: null,
-      users: new Set<number>(),
-      admin: new Set<number>(),
-      muted: new Map<number, Date>(),
-      banned: new Set<number>(),
-    };
-    this.channelMap.set(newChannel.id, newChannel);
 
-    return newChannel;
-  }
 
-  joinChannel(user: ChannelUser, channel: Channel) {
-    user.socket.join(String(channel.id));
-    user.joined.add(channel.id);
-    channel.users.add(user.id);
-  }
+
+
+
+
+
+  // Controller 
 
   create(userId: number, data: CreateChannelDto) {
     const createdBy: ChannelUser = this.getUser(userId);
-    const newChannel: Channel = this.createChannel(data);
+    const newChannel: Channel = this.channelClass.createChannel(data);
 
     newChannel.owner = createdBy.id;
-    this.joinChannel(createdBy, newChannel);
+    this.channelClass.joinChannel(createdBy, newChannel);
 
     // socket message
     createdBy.socket.emit('message', {
@@ -285,7 +299,7 @@ export class ChannelsService {
   }
 
   join(userId: number, channelId: number, password: string) {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const user: ChannelUser = this.getUser(userId);
 
     if (channel.password && password != channel.password) {
@@ -303,26 +317,19 @@ export class ChannelsService {
         code: HttpStatus.FORBIDDEN,
         message: 'You are connot join the channel.',
       };
-    } else if (this.checkIsJoined(channel, user.id)) {
+    } else if (this.channelClass.checkIsJoined(channel, user.id)) {
       throw {
         code: HttpStatus.CONFLICT,
         message: 'This channel is already participating.',
       };
-    } else if (channel.isPrivate && !user.invited.has(channel.id)) {
-      throw {
-        code: HttpStatus.FORBIDDEN,
-        message: 'You are connot join the channel.',
-      };
-    } else if (user.invited.has(channel.id)) {
-      user.invited.delete(channel.id);
     }
 
-    this.joinChannel(user, channel);
+    this.channelClass.joinChannel(user, channel);
     this.noticeToChannel(channelId, `${user.name} 님이 채널에 참여하였습니다.`);
   }
 
   leave(userId: number, channelId: number) {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const user: ChannelUser = this.getUserJoinedChannel(userId, channelId);
 
     if (channel.owner == user.id) {
@@ -337,31 +344,9 @@ export class ChannelsService {
     this.noticeToChannel(channelId, `${user.name} 님이 나가셨습니다.`);
   }
 
-  checkCanListed(channel: Channel, userId: number): boolean {
-    const isJoined = channel.users.has(userId);
-    const isPrivate = channel.isPrivate;
-    const isDm = channel.isDm;
-
-    return ((isPrivate || isDm) && isJoined) || !isPrivate;
-  }
-
-  checkIsJoined(channel: Channel, userId: number): boolean {
-    return channel.users.has(userId);
-  }
-
-  checkChannelAthor
-
-  checkListedRange(query, channel: Channel) {
-    return (
-      !(query.isPublic && !channel.isPrivate && !channel.isDm) &&
-      !(query.isPriv && channel.isPrivate && !channel.isDm) &&
-      !(query.isDm && !channel.isPrivate && channel.isDm)
-    );
-  }
-
   list(userId: number, query) {
     const data = [];
-    const channels = this.getChannelValues();
+    const channels = this.channelClass.getAll();
 
     if (!query.isPublic && !query.isPriv && !query.isDm) {
       query.isPublic = true;
@@ -369,9 +354,9 @@ export class ChannelsService {
 
     channels.forEach((channel) => {
       if (
-        !this.checkCanListed(channel, userId) ||
-        (query.isEnter && !this.checkIsJoined(channel, userId)) ||
-        this.checkListedRange(query, channel)
+        !this.channelClass.checkCanListed(channel, userId) ||
+        (query.isEnter && !this.channelClass.checkIsJoined(channel, userId)) ||
+        this.channelClass.checkListedRange(channel, query)
       ) {
         return;
       }
@@ -392,7 +377,7 @@ export class ChannelsService {
         channel.users.forEach((id) => {
           if (id != userId) {
             info.dmUserId = id;
-            info.title = this.channelUserMap.get(id).name;
+            info.title = this.getUser(id).name;
           }
         });
       }
@@ -402,23 +387,10 @@ export class ChannelsService {
     return data;
   }
 
-  checkCanGetInfo(channel: Channel, userId: number): boolean {
-    const isJoined = channel.users.has(userId);
-    const isPrivate = channel.isPrivate;
-    const isDm = channel.isDm;
-    const isPassword = channel.password;
-
-    return (
-      ((isPrivate || isDm) && isJoined) ||
-      (!isPrivate && isPassword && !isDm && isJoined) ||
-      (!isPrivate && !isPassword && !isDm)
-    );
-  }
-
   getChannelInfo(userId: number, channelId: number) {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
 
-    if (!this.checkCanGetInfo(channel, userId)) {
+    if (!this.channelClass.checkCanGetInfo(channel, userId)) {
       throw {
         code: HttpStatus.FORBIDDEN,
         message: 'You are not authorized to this channel.',
@@ -435,7 +407,7 @@ export class ChannelsService {
         role: role,
         isMuted: channel.muted.has(id),
       })
-    })
+    });
 
     const channelInfo = {
       id: channel.id,
@@ -454,7 +426,7 @@ export class ChannelsService {
   }
 
   setChannelInfo(userId: number, channelId: number, data) {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const settedBy: ChannelUser = this.getUser(userId);
 
     if (channel.isDm) {
@@ -488,7 +460,7 @@ export class ChannelsService {
     settedUserId: number,
     role: string
   ) {
-    const channel: Channel = this.getChannel(channelId);
+    const channel: Channel = this.channelClass.get(channelId);
     const settedBy: ChannelUser = this.getUserJoinedChannel(userId, channelId);
     const settedUser: ChannelUser = this.getUserJoinedChannel(settedUserId, channelId);
 
@@ -518,64 +490,35 @@ export class ChannelsService {
     }
   }
 
-  getDms(user: ChannelUser): Channel[] {
-    const data = [];
-
-    user.joined.forEach((channelId) => {
-      const channel: Channel = this.channelMap.get(channelId);
-      if (channel && channel.isDm) {
-        data.push(channel);
-      }
-    });
-
-    return data;
-  }
-
   initDirectMessage(userId: number, invitedUserId: number) {
-    const invitedUser: ChannelUser = this.channelUserMap.get(invitedUserId);
-    const user: ChannelUser = this.channelUserMap.get(userId);
+    const invitedUser: ChannelUser = this.getUser(invitedUserId);
+    const user: ChannelUser = this.getUser(userId);
 
-    if (!user || !invitedUser) {
-      throw {
-        code: HttpStatus.BAD_REQUEST,
-        message: 'This user does not exist.',
-      };
-    } else if (invitedUser.blockUser.has(userId)) {
+    if (invitedUser.blockUser.has(userId)) {
       throw {
         code: HttpStatus.BAD_REQUEST,
         message: 'You cannot send dm to.',
       };
     }
 
-    const dms: Channel[] = this.getDms(user);
+    const dms: Channel[] = this.channelClass.getAllDms(user);
     const found = dms.find(
       (channel) =>
         channel.users.has(user.id) && channel.users.has(invitedUserId)
     );
 
     if (!found) {
-      const newChannel = this.createChannel({ isDm: true });
-      this.joinChannel(user, newChannel);
-      this.joinChannel(invitedUser, newChannel);
+      const newChannel = this.channelClass.createChannel({ isDm: true });
+      this.channelClass.joinChannel(user, newChannel);
+      this.channelClass.joinChannel(invitedUser, newChannel);
       return { id: newChannel.id };
     }
 
     return { id: found.id };
   }
 
-  names(socket: Socket, channelId: number) {
-    const channel = this.getChannel(channelId);
-    if (!channel) {
-      socket.emit('error', { code: 4, msg: 'This channel does not exist.' });
-      return;
-    }
-
-    const data = this.getUsernameList();
-    socket.emit('names', data);
-  }
-
   invite(userId: number, channelId: number, invitedUserId: number[]) {
-    const channel = this.getChannel(channelId);
+    const channel = this.channelClass.get(channelId);
     const invitedBy = this.getUserJoinedChannel(userId, channelId);
     const role = this.getChannelUserRole(channel.id, invitedBy.id);
     
@@ -594,15 +537,14 @@ export class ChannelsService {
     invitedUserId.forEach((id) => {
       const invitedUser = this.getUser(id);
 
-      if (this.checkIsJoined(channel, invitedUser.id)) {
+      if (this.channelClass.checkIsJoined(channel, invitedUser.id)) {
         throw {
           code: HttpStatus.CONFLICT,
           message: 'The user is already joining the channel.',
         };
       }
   
-      invitedUser.invited.add(channel.id);
-      this.joinChannel(invitedUser, channel);
+      this.channelClass.joinChannel(invitedUser, channel);
       
       // socket massage
       invitedUser.socket.emit('invited', { channelId: channel.id });
@@ -611,36 +553,11 @@ export class ChannelsService {
   }
 
   setUserStatus(userId: number, channelId: number, settedUserId: number, status: string) {
-    const channel: Channel = this.channelMap.get(channelId);
-    const user: ChannelUser = this.getUser(userId);
-    const settedUser: ChannelUser = this.getUser(settedUserId);
+    const channel: Channel = this.channelClass.get(channelId);
+    const user: ChannelUser = this.getUserJoinedChannel(userId, channelId);
+    const settedUser: ChannelUser = this.getUserJoinedChannel(settedUserId, channelId);
 
-    if (!channel) {
-      throw {
-        code: HttpStatus.BAD_REQUEST,
-        message: 'This channel does not exist.',
-      };
-    } else if (!user) {
-      throw {
-        code: HttpStatus.BAD_REQUEST,
-        message: 'This user does not exist.',
-      };
-    } else if (!settedUser) {
-      throw {
-        code: HttpStatus.BAD_REQUEST,
-        message: 'This user does not exist.',
-      };
-    } else if (!channel.users.has(user.id)) {
-      throw {
-        code: HttpStatus.BAD_REQUEST,
-        message: 'This user is not in the channel.',
-      };
-    } else if (!channel.users.has(settedUser.id)) {
-      throw {
-        code: HttpStatus.BAD_REQUEST,
-        message: 'This user is not in the channel.',
-      };
-    } else if (channel.owner != user.id && !channel.admin.has(user.id)) {
+    if (channel.owner != user.id && !channel.admin.has(user.id)) {
       throw {
         code: HttpStatus.UNAUTHORIZED,
         message: 'This user does not have permission to do this.',
@@ -659,31 +576,4 @@ export class ChannelsService {
     }
   }
 
-  kick(user: ChannelUser, channel: Channel, kickedUser: ChannelUser) {
-    this.setUserLeave(kickedUser, channel);
-    this.noticeToChannel(
-      channel.id,
-      `${user.name}님이 ${kickedUser.name} 님을 채널에서 강퇴하였습니다.`
-    );
-  }
-
-  ban(user: ChannelUser, channel: Channel, bannedUser: ChannelUser) {
-    this.setUserLeave(bannedUser, channel);
-    channel.banned.add(bannedUser.id);
-    this.noticeToChannel(
-      channel.id,
-      `${user.name}님이 ${bannedUser.name} 님을 채널에서 차단하였습니다.`
-    );
-  }
-
-  mute(user: ChannelUser, channel: Channel, mutedUser: ChannelUser, seconds: number) {
-    const now = new Date();
-    const expiresAt = new Date(now.getTime() + (ONESECOND * seconds));
-
-    channel.muted.set(mutedUser.id, expiresAt);
-    this.noticeToChannel(
-      channel.id,
-      `${user.name}님이 ${mutedUser.name} 님을 채널에서 조용히 시켰습니다.`
-    );
-  }
 }
